# apispec.yaml
openapi: 3.0.0
info:
  title: Portfolio
  version: 1.0.0
servers:
  - url: http://localhost:3000
tags:
  - name: simple
  - name: to_do_app
paths:
  /simple/add-two-number:
    post:
      tags:
        - simple
      description: Add 2 number
      summary: Add 2 number
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputAdd2Number'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponse_add_2_number'
        '400':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'

  /to-do:
    get:
      tags:
        - to_do_app
      description: get all list
      summary: get all list 
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_get_all_toDo'
        '400':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
    post:
      tags:
        - to_do_app
      description: Add to do list
      summary: Add to do list 
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputAddTodo'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response_rows_Affected'
        '400':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
  
  /to-do/{id}:
    put:
      tags:
        - to_do_app
      description: Update to do list by id
      summary: Update to do list by id 
      parameters:
        - name: id
          in: path
          description: id to_do list
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputUpdateTodo'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response_rows_Affected'
        '400':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
    
    delete:
      tags:
        - to_do_app
      description: Delete to do list by id
      summary: Delete to do list by id 
      parameters:
        - name: id
          in: path
          description: id to_do list
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response_rows_Affected'
        '400':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'

components:
  # securitySchemes:
  #   CategoryAuth:
  #     type: apiKey
  #     in: header
  #     name: X-API-Key
  #     description: Authentication for Category API
  schemas:
    InputAdd2Number:
      type: object
      properties:
        number_1:
          type: number
        number_2:
          type: number
    
    InputAddTodo:
      type: object
      properties:
        title:
          type: string
      additionalProperties: false
      required:
        - title
    
    InputUpdateTodo:
      type: object
      properties:
        title:
          type: string
        status:
          type: boolean
      additionalProperties: false
      required:
        - title

    WebResponse_add_2_number:
      type: object
      properties:
        status:
          type: string
        data:
          type: number
      additionalProperties: false
      required:
        - status
        - data

    ErrorBadRequest:
      type: object
      properties:
        status:
          type: string
        data:
          type: string # null
        message:
          type: string
      additionalProperties: false
      required:
        - status
        - data
        - message

    Response_rows_Affected:
      type: object
      properties:
        status:
          type: string
        data:
          type: object
          properties:
            rows_affected:
              type: integer
          required:
            - rows_affected
          additionalProperties: false
      additionalProperties: false
      required:
        - status
        - data
    response_get_all_toDo:
      type: object
      properties:
        status:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              title:
                type: string
              status:
                type: boolean
              created_at:
                type: integer
              updated_at:
                type: integer
            required:
              - id
              - title
              - status
              - created_at
              - updated_at
      additionalProperties: false
      required:
        - status
        - data